Fix clang warnings for linux builds (#6140)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 274beb9dd42dc6c7d7dd3ff9415a41e1393f0133

+++ starboard/nplb/posix_compliance/posix_directory_create_test.cc
@@ -77,7 +77,7 @@ TEST(PosixDirectoryCreateTest, SunnyDayTempDirectoryManySeparators) {
   ASSERT_TRUE(system_path_success);
   const int new_size = starboard::strlcat(
       temp_path.data(), kManyFileSeparators.c_str(), kSbFileMaxPath);
-  ASSERT_LT(new_size, kSbFileMaxPath);
+  ASSERT_LT(new_size, static_cast<int>(kSbFileMaxPath));
 
   struct stat info;
   EXPECT_TRUE(stat(temp_path.data(), &info) == 0 && S_ISDIR(info.st_mode));
